forgot config

+++ conf/config.py
@@ -16,5 +16,5 @@ bnb_config = BitsAndBytesConfig(
     load_in_4bit=True,
     bnb_4bit_use_double_quant=True,
     bnb_4bit_quant_type="nf4",
-    bnb_4bit_compute_dtype=torch.float32,  # ← this is the key change
+    bnb_4bit_compute_dtype=torch.float16,
 )
